🔥fix: Change data types from decimal to integer for specific fields in catatan_partograf migration

+++ database/migrations/2025_11_11_071100_create_catatan_partograf_table.php
@@ -22,13 +22,13 @@ return new class extends Migration
         Schema::create('catatan_partograf', function (Blueprint $table) {
             $table->string('id', 25)->primary();
             $table->string('waktu_catat', 25)->nullable();
-            $table->decimal('djj')->nullable();
-            $table->decimal('pembukaan_servik')->nullable();
-            $table->decimal('penurunan_kepala')->nullable();
-            $table->decimal('nadi_ibu')->nullable();
+            $table->integer('djj')->nullable();
+            $table->integer('pembukaan_servik')->nullable();
+            $table->integer('penurunan_kepala')->nullable();
+            $table->integer('nadi_ibu')->nullable();
             $table->decimal('suhu_ibu')->nullable();
-            $table->decimal('sistolik')->nullable();
-            $table->decimal('diastolik')->nullable();
+            $table->integer('sistolik')->nullable();
+            $table->integer('diastolik')->nullable();
             $table->enum('aseton', ['-', '+'])->nullable();  // Laravel enum tetap bisa
             $table->enum('protein', ['-', '+', '++', '+++'])->nullable();
             $table->decimal('volume_urine')->nullable();
